Avance Viernes Biblioteca por la mañana.

+++ ClaseMiPrimerAPI/Model/RequestVendedor.cs
@@ -4,6 +4,6 @@
     {
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public float Salario { get; set; }
+        public int Salario { get; set; }
     }
 }
